// app/api/(@types)/SearchOutcome.ts
/**
 * The SearchOutcome type definition.
 *
 * @file SearchOutcome.ts
 * @summary SearchOutcome type definition
 * @description A type definition for the search outcome
 */
export type SearchOutcome = {
  id: string;
  name: string;
  price: string;
  brand: string;
  category: string;
  position: string;
  variant: string;
  list: string;
  unit_sale_price: string;
  stock: string;
  product_image_url: string;
};